setup for de detail page
