+++ src/lib/Pages/Home/Title.svelte
@@ -1,8 +1,11 @@
 <script>
   import { MonthlyNavigation } from "$lib";
+  import { page } from "$app/stores";
   export let company;
   export let nextMonth;
   export let previousMonth;
+
+  let currentPage = $page.url.pathname;
 </script>
 
 <section>
@@ -10,7 +13,9 @@
     <h1>{company[0].title}/home</h1>
     <p>Automatische Scan</p>
   </div>
-  <MonthlyNavigation onNextMonth={nextMonth} onPreviousMonth={previousMonth} />
+  {#if currentPage === "/"}
+    <MonthlyNavigation onNextMonth={nextMonth} onPreviousMonth={previousMonth} />
+  {/if}
 </section>
 
 <style>
